requirements.txt
./wheels/ndi_python-5.1.1.5-cp311-cp311-linux_x86_64.whl
blinker==1.7.0
Bootstrap-Flask==2.4.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
PyYAML==6.0.1
Werkzeug==3.0.2
WTForms==3.1.2
psutil==5.9.8